workflows: build-only for x86_64 and aarch64
Provide build-only workflow which executes minimal 'make && make check
&& make test' for every supported architecture (namely, x86_64/amd64 and
aarch64/arm64). This workflow is triggered for every push on any
non-private branch (that is, branch-name does not begin with '_') in
order to ensure that build process did not break. In particular, that
all build-tools work properly on architectures other than the one used
by developer.

Signed-off-by: Shachar Sharon <ssharon@redhat.com>